+++ src/application/domain/Account.ts
@@ -1,7 +1,9 @@
 import deps, { IDependencies } from '../../infra/dependencies'
 import { IAccount, ICurrency } from '../types';
+import { Currency } from './Currency';
 
 type inAccount = {
+  account_id?: number | null,
   name: string,
   currency_iso_code: string,
   description?: string | null,
@@ -20,7 +22,7 @@ export class Account {
   ) {
     this.data = {
       ...data,
-      account_id: null,
+      account_id: data.account_id || null,
       description: data.description || null,
       bank_name: data.bank_name || null,
       starting_balance: data.starting_balance || 0,
@@ -39,6 +41,11 @@ export class Account {
     return this.data;
   }
 
+  public checkIsPersisted() {
+    if (!this.data.account_id) throw new Error(`Account ${this.data.name} is not persisted`);
+    return true;
+  }
+
   public async persist(): Promise<IAccount> {
     this.data = await this.deps.repositories.account.create(this.data);
     return this.data;;
@@ -49,18 +56,6 @@ export class Account {
     // TODO: Should be static but I need those dependencies
     const accounts: IAccount[] = await this.deps.db.table('accounts').select('*');
     return accounts;
-  }
-
-  public async getBalance() {
-    if (!this.data.account_id) throw new Error('Account is not persisted');
-    const latest = await this.deps.repositories.account
-      .getLastBalanceUpdate(this.data.account_id);
-
-    if (latest) {
-      return latest.new_balance;
-    } else {
-      return this.data.starting_balance;
-    }
   }
 
   public async updateBalance({
@@ -81,10 +76,89 @@ export class Account {
     });
   }
 
-}
+  public async transferMoney({
+    to_account,
+    amount,
+    open_at,
+    description = null,
+    fees = 0
+  }: {
+    to_account: Account,
+    amount: number,
+    open_at: Date,
+    description?: string | null;
+    fees?: number;
+  }) {
+    if (!this.data.account_id) throw new Error('Account is not persisted');
+    if (!to_account.data.account_id) throw new Error('to_account is not persisted');
 
-export class AccountMovement {
+    const transfer = await this.deps.repositories.accountActivity.create({
+      account_withdrawal_id: this.data.account_id,
+      account_deposit_id: to_account.data.account_id,
+      amount_withdrawal: -amount,
+      amount_deposit: amount,
+      fees,
+      description,
+      open_at,
+    });
 
-  constructor() { }
+    return transfer;
+  }
+
+  public async getBalance(currenct?: Currency) {
+    if (!this.data.account_id) throw new Error('Account is not persisted');
+
+    const [
+      deposits,
+      withdrawals
+    ] = await Promise.all([
+      this.deps.repositories.accountActivity.listDeposits(this.data.account_id),
+      this.deps.repositories.accountActivity.listWithdrawals(this.data.account_id),
+    ]);
+
+    const totalDeposits = deposits.reduce((prev, curr) => {
+      return prev + curr.amount_deposit;
+    }, 0);
+
+    const totalWithdrawal = withdrawals.reduce((prev, curr) => {
+      return prev + curr.amount_deposit;
+    }, 0);
+
+    const balance = this.data.starting_balance + totalDeposits - totalWithdrawal;
+    return balance;
+  }
+
+  public async getStats() {
+
+    if (!this.data.account_id) throw new Error('Account is not persisted');
+
+    const [
+      deposits,
+      withdrawals
+    ] = await Promise.all([
+      this.deps.repositories.accountActivity.listDeposits(this.data.account_id),
+      this.deps.repositories.accountActivity.listWithdrawals(this.data.account_id),
+    ]);
+
+    const totalDeposits = deposits.reduce((prev, curr) => {
+      return prev + curr.amount_deposit;
+    }, 0);
+
+    const totalWithdrawal = withdrawals.reduce((prev, curr) => {
+      return prev + curr.amount_deposit;
+    }, 0);
+
+    console.info('Account All Information')
+    console.info(this.data);
+
+    console.info('List deposits');
+    console.table(deposits);
+
+    console.info('List withdrawals');
+    console.table(withdrawals);
+
+    const balance = this.data.starting_balance + totalDeposits - totalWithdrawal;
+    console.info(`Current balance is ${balance} ${this.data.currency_iso_code}`)
+  }
 
 }