+++ src/infra/dependencies.ts
@@ -1,4 +1,5 @@
 import { Knex } from "knex";
+import { AccountActivityRepo } from "../application/repositories/AccountActivityRepo";
 import { AccountRepo } from "../application/repositories/AccountRepo";
 import { CurrencyRepo } from "../application/repositories/CurrencyRepo";
 import { db } from "./database";
@@ -9,17 +10,20 @@ export interface IDependencies {
   repositories: {
     account: AccountRepo,
     currency: CurrencyRepo,
+    accountActivity: AccountActivityRepo
   }
 }
 
 const currencyRepo = new CurrencyRepo(db);
 const accountRepo = new AccountRepo(db, currencyRepo);
+const accountActivityRepo = new AccountActivityRepo(db, currencyRepo);
 
 const deps: IDependencies = {
   db,
   repositories: {
     account: accountRepo,
     currency: currencyRepo,
+    accountActivity: accountActivityRepo
   }
 }